requirements.txt
pillow
schedule
textwrap3

